requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
passlib
jwt
PyJWT
config
pydantic
pydantic[email]
python-multipart